requirements.txt

Flask==3.0.3
mysql-connector-python==8.0.33
Werkzeug==3.0.3
python-dotenv==1.0.1

gunicorn==20.1.0
